Fixed style of example
While it's certainly nice that C++ allows you to not have {} to
encapsulate a function body if that body only consists of a single
statement (here: `try{…}`), it is not of any familiarity to C++ coders
to omit them.

Same goes for putting `main`'s argc/argv arguments on multiple lines.

Also, defining a macro PRIVATE_SOMETHING just to use it once one line of
code below is a bit of a strange thing to do.

So, I got rid of the macro and applied the clang-format style from
https://github.com/boost-experimental/vc